backend/pyproject.toml
[project]
name = "cv-modifier-saas"
version = "0.1.0"
description = "SaaS pour adapter automatiquement CV et lettres de motivation aux offres d'emploi"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "fastapi-users[sqlalchemy]>=14.0.0",
    "anthropic>=0.87.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.13.0",
    "fpdf2>=2.8.7",
    "python-dotenv>=1.1.0",
    "pydantic-settings>=2.7.0",
    "pypdf>=5.0.0",
    "playwright>=1.49.0",
    "python-multipart>=0.0.18",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
prod = [
    "asyncpg>=0.30.0",
    "boto3>=1.35.0",
    "stripe>=12.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
